fix bug & change some controller (Notification & Friend)

+++ src/main/java/com/example/backend/repository/CallHistoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.backend.model.entity.CallHistory;
 
-public interface CallHistoryRepository extends JpaRepository<CallHistory, Integer> {
+public interface CallHistoryRepository extends JpaRepository<CallHistory, Long> {
 
 }
