+++ src/main/java/com/example/backend/repository/CallRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.backend.model.entity.Call;
 
-public interface CallRepository extends JpaRepository<Call, Integer> {
+public interface CallRepository extends JpaRepository<Call, Long> {
 
 }
