+++ src/main/java/com/example/backend/repository/NotificationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.backend.model.entity.Notification;
 
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
     List<Notification> findByReceiverIdOrderByCreatedAtDesc(Long receiverId);
 }
